Add files via upload
Recognizes the pool at Lindbergh High School and identifies which part of the pool is visible on the animated pool window. The animated pool window is not reliable because the pool corners jump around. This is code before adding new way of locating pool corners with Kalman filter and optical flow.