Add files via upload
First working version of the DATA COLLECTION and PRESENTATION main functions. In the DATA COLLECTION mode, the user can input the position of the players, ball, and sides of the pool, which are then saved in a file. In the PRESENTATION mode, the program reads in the data from the saved file and displays the positions on an animated pool window as well as in the camera window.