Add files via upload
The new pool identifier class successfully integrated with the actual program. Next steps are to 1) research how to incorporate a computer vision recognition, 2) make the program optimize the choice of the line so it doesn’t have to go through points but can go through a line of best fit, 3) enable the program to operate with more than 9 points, and 4) add compatibility for using the positions of the goal line, 2M, 5M, 6M, half, and goal posts.